Modified structure and docs of radio filter.
