+++ src/Filter/modules/RadioFilter/index.ts
@@ -1,0 +1,3 @@
+import RadioFilter from './RadioFilter';
+
+export default RadioFilter;
