+++ src/Filter/types/index.ts
@@ -301,6 +301,20 @@ export interface FilterContainerProps {
 };
 
 /**
+ * `RadioFilterProps` are the props required to be supplied as the first argument of
+ * the RadioFilter component.
+ */
+export interface RadioFilterProps {
+    initialValue: string;
+    defaultValue: string;
+    options: {
+        label: string;
+        value: string;
+    }[];
+    label: string;
+}
+
+/**
  * `RadioFilterOptions` is any valid `FilterModule` property (excluding description and label)
  * meant to override default text filter behaviour.
  */
