+++ src/stories/Filter/RadioFilter/RadioFilterDocs.tsx
@@ -1,26 +1,12 @@
 import { FC } from 'react';
-import { RadioFilterOptions } from 'src/Filter/types';
+import { RadioFilterOptions, RadioFilterProps } from 'src/Filter/types';
 
 export interface RadioFilterArgs {
     /**
-     * The label to display for the radio filter component.
+     * The props required to be supplied as the first argument of
+     * the RadioFilter component.
      */
-    label: string;
-    /**
-     * The value the radio filter starts with.
-     */
-    initialValue: string;
-    /**
-     * The default value when there is no value selected.
-     */
-    defaultValue: string;
-    /**
-     * The list of options to choose from in the radio filter.
-     */
-    options: {
-        label: string;
-        value: string;
-    }[];
+    radioFilterProps: RadioFilterProps;
     /**
      * Any valid `FilterModule` property (excluding description and label)
      * which will override default radio filter behaviour.
@@ -30,7 +16,7 @@ export interface RadioFilterArgs {
 
 /**
  * RadioFilter Component
- * 
+ *
  * The RadioFilter component is a filter by use of radio group. While the default
  * behaviour should suffice, any valid `FilterModule` property (excluding label) can
  * be supplied via the `radioFilterOptions` parameter to change how the filter looks and acts.
